Fix memory leaks found with Valgrind; compile with --std=c++11
There are two kinds of problems. Uninitialized values, and leaks. I am ignoring leaks for now as most memory is held until VM quits and the OS can free it just as well.

I started setting threads as detached, so they do not wait around to be joined. This still leaks, because the VM quits before the threads have a chance to gracefully shut down. But it is again OK.

Uninitialized values are a hard problem, because they come from the assembly and they are hard to trace. They crop up at seemingly important spots, yet the VM runs fine. Either Memcheck is somehow wrong (I am not prepared to rule it out just yet) or the assembly is subtly wrong (more likely).

==25528== 8 bytes in 2 blocks are definitely lost in loss record 2 of 80 // FIXED
==25528==    at 0x402B41C: malloc (in /usr/lib/valgrind/vgpreload_memcheck-x86-linux.so)
==25528==    by 0x82CDAA6: mainWrapper(void*) (in /home/jirka/Bin/jirkadanekStrongtalk/cmakebuild/strongtalk-test)
==25528==    by 0x408A314: start_thread (in /usr/lib32/libpthread-2.22.so)
==25528==    by 0x4377D6D: clone (in /usr/lib32/libc-2.22.so)

==25528== 144 bytes in 1 blocks are possibly lost in loss record 46 of 80 // hopefully FIXED
==25528==    at 0x402C1F8: calloc (in /usr/lib/valgrind/vgpreload_memcheck-x86-linux.so)
==25528==    by 0x40111D6: allocate_dtv (in /usr/lib32/ld-2.22.so)
==25528==    by 0x4011C1B: _dl_allocate_tls (in /usr/lib32/ld-2.22.so)
==25528==    by 0x408AD23: pthread_create@@GLIBC_2.1 (in /usr/lib32/libpthread-2.22.so)
==25528==    by 0x82CE567: os_init() (in /home/jirka/Bin/jirkadanekStrongtalk/cmakebuild/strongtalk-test)